@@ -11,131 +11,153 @@
   "bugs": "https://github.com/gilbarbara/react-redux-saga-boilerplate/issues",
   "license": "MIT",
   "dependencies": {
-    "bootstrap": "^4.0.0-alpha",
-    "classlist-polyfill": "^1.0",
-    "classnames": "^2.2",
-    "core-decorators": "^0.15",
-    "core-js": "^2.4",
-    "ease-component": "^1.0",
-    "fbjs": "^0.8",
-    "history": "^3.2",
-    "isomorphic-fetch": "^2.2",
-    "lodash": "^4.17",
-    "mobile-detect": "^1.3",
-    "modernizr": "^3.3",
-    "moment": "^2.17",
-    "react": "^15.4",
-    "react-dom": "^15.4",
-    "react-inlinesvg": "^0.5",
-    "react-redux": "^5.0",
-    "react-router": "^3.0",
-    "react-router-redux": "^4.0",
-    "redux": "^3.6",
-    "redux-persist": "^4.3",
-    "redux-saga": "^0.14",
-    "redux-thunk": "^2.2",
-    "sass-mediaqueries": "^1.6",
-    "scroll": "^2.0",
-    "scroll-doc": "^0.1"
+    "bootstrap": "^4.0.0-beta.2",
+    "classlist-polyfill": "^1.2.0",
+    "classnames": "^2.2.5",
+    "core-js": "^2.5.1",
+    "ease-component": "^1.0.0",
+    "fbjs": "^0.8.16",
+    "history": "^4.7.2",
+    "immutability-helper": "^2.4.0",
+    "isomorphic-fetch": "^2.2.1",
+    "lodash": "^4.17.4",
+    "mobile-detect": "^1.3.7",
+    "modernizr": "^3.5.0",
+    "prop-types": "^15.6.0",
+    "qs": "^6.5.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-helmet": "^5.2.0",
+    "react-inlinesvg": "^0.7.0",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.8",
+    "redux": "^3.7.2",
+    "redux-persist": "^5.2.2",
+    "redux-saga": "^0.16.0",
+    "redux-thunk": "^2.2.0",
+    "sass-mediaqueries": "^1.6.1",
+    "scroll": "^2.0.1",
+    "scroll-doc": "^0.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7",
-    "babel-core": "^6.22",
-    "babel-eslint": "^7.1",
-    "babel-jest": "^18.0",
-    "babel-loader": "^6.2",
-    "babel-plugin-istanbul": "^4.0",
-    "babel-plugin-lodash": "^3.2",
-    "babel-plugin-transform-decorators-legacy": "^1.3",
-    "babel-plugin-transform-object-assign": "^6.22",
-    "babel-plugin-transform-runtime": "^6.22",
-    "babel-preset-latest": "^6.22",
-    "babel-preset-react": "^6.22",
-    "babel-preset-stage-1": "^6.22",
-    "babel-register": "^6.22",
-    "babel-runtime": "^6.22",
-    "browser-sync": "^2.18",
-    "browser-sync-webpack-plugin": "^1.1",
-    "chalk": "^1.1",
-    "chromedriver": "^2.27",
-    "clean-webpack-plugin": "^0.1",
-    "copy-webpack-plugin": "^4.0",
-    "css-loader": "^0.26",
-    "del": "^2.2",
-    "enzyme": "^2.7",
-    "esdoc": "^0.5",
+    "autoprefixer": "^7.1.6",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.1",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-array-includes": "^2.0.3",
+    "babel-plugin-istanbul": "^4.1.5",
+    "babel-plugin-lodash": "^3.2.11",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-1": "^6.24.1",
+    "babel-runtime": "^6.26.0",
+    "browser-sync": "^2.18.13",
+    "browser-sync-webpack-plugin": "^1.2.0",
+    "bundlesize": "^0.15.3",
+    "case-sensitive-paths-webpack-plugin": "^2.1.1",
+    "chalk": "^2.3.0",
+    "chromedriver": "^2.33.2",
+    "clean-webpack-plugin": "^0.1.17",
+    "copy-webpack-plugin": "^4.2.0",
+    "css-loader": "^0.28.7",
+    "date-fns": "^1.29.0",
+    "del": "^3.0.0",
+    "dotenv": "^4.0.0",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.2",
+    "esdoc": "^1.0.3",
     "esdoc-es7-plugin": "^0.0.3",
-    "eslint": "^3.15",
-    "eslint-config-airbnb": "^14.1",
-    "eslint-import-resolver-webpack": "^0.8",
-    "eslint-loader": "^1.6",
-    "eslint-plugin-babel": "^4.0",
-    "eslint-plugin-flowtype": "^2.30",
-    "eslint-plugin-import": "^2.2",
-    "eslint-plugin-jsx-a11y": "^4.0",
-    "eslint-plugin-react": "^6.9",
-    "exports-loader": "^0.6",
-    "expose-loader": "^0.7",
-    "extract-text-webpack-plugin": "^2.0.0-beta",
-    "fetch-mock": "^5.9",
-    "file-loader": "^0.10",
-    "gh-pages": "^0.12",
-    "html-loader": "^0.4",
-    "html-webpack-plugin": "^2.28",
-    "husky": "^0.13",
-    "image-webpack-loader": "^3.2",
-    "imports-loader": "^0.7",
-    "jest-cli": "^18.1",
-    "jsdom": "^9.10",
-    "json-loader": "^0.5",
-    "markdown-loader": "^2.0",
-    "modernizr-loader": "^1.0",
-    "nightwatch": "^0.9",
-    "node-sass": "^4.5",
-    "null-loader": "^0.1",
-    "offline-plugin": "^4.6",
-    "postcss-loader": "^1.2",
-    "raw-loader": "^0.5",
-    "react-addons-test-utils": "^15.4",
-    "react-hot-loader": "^3.0.0-beta",
-    "reactotron-react-js": "^1.7",
-    "reactotron-redux": "^1.7",
-    "redux-diff-logger": "^0.0.9",
-    "redux-logger": "^2.8",
-    "redux-mock-store": "^1.2",
-    "sass-loader": "^5.0",
-    "script-loader": "^0.7",
-    "selenium-server-standalone-jar": "^3.0",
-    "source-map-support": "^0.4",
-    "style-loader": "^0.13",
-    "stylelint": "^7.8",
-    "stylelint-config-standard": "^16.0",
-    "stylelint-order": "^0.2",
-    "stylelint-scss": "^1.4",
-    "url-loader": "^0.5",
-    "webpack": "^2.2",
-    "webpack-dev-server": "^2.3",
-    "webpack-merge": "^2.6"
+    "eslint": "^4.10.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-import-resolver-webpack": "^0.8.3",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-flowtype": "^2.39.1",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.4.0",
+    "expose-loader": "^0.7.3",
+    "extract-text-webpack-plugin": "^3.0.2",
+    "fetch-mock": "^5.13.1",
+    "file-loader": "^1.1.5",
+    "gh-pages": "^1.0.0",
+    "git-info-webpack-plugin": "^1.0.1",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
+    "husky": "^0.14.3",
+    "image-webpack-loader": "^3.4.2",
+    "jest": "^21.2.1",
+    "jest-enzyme": "^4.0.1",
+    "jsdom": "^11.3.0",
+    "json-loader": "^0.5.7",
+    "markdown-loader": "^2.0.1",
+    "modernizr-loader": "^1.0.1",
+    "nightwatch": "^0.9.16",
+    "node-sass": "^4.5.3",
+    "offline-plugin": "^4.8.4",
+    "postcss-loader": "^2.0.8",
+    "raw-loader": "^0.5.1",
+    "react-addons-test-utils": "^15.6.2",
+    "react-dev-utils": "^4.2.1",
+    "react-error-overlay": "^3.0.0",
+    "react-hot-loader": "^3.1.1",
+    "redux-logger": "^3.0.6",
+    "redux-mock-store": "^1.3.0",
+    "sass-loader": "^6.0.6",
+    "script-loader": "^0.7.2",
+    "selenium-server-standalone-jar": "^3.6.0",
+    "source-map-support": "^0.5.0",
+    "style-loader": "^0.19.0",
+    "stylelint": "^8.2.0",
+    "stylelint-config-standard": "^17.0.0",
+    "stylelint-declaration-strict-value": "^1.0.4",
+    "stylelint-order": "^0.7.0",
+    "stylelint-scss": "^2.1.0",
+    "url-loader": "^0.6.2",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.3",
+    "webpack-merge": "^4.1.1"
   },
   "scripts": {
-    "start": "node config/webpack.server.js",
-    "build": "NODE_ENV=production webpack --config config/webpack.prod.js",
-    "build:pages": "NODE_ENV=production webpack --config config/webpack.prod.js --env.target=pages",
+    "start": "APP_ENV=dev node tools/start.js",
+    "build": "NODE_ENV=production APP_ENV=production webpack --config config/webpack.config.prod.js",
+    "build:pages": "NODE_ENV=production APP_ENV=pages webpack --config config/webpack.config.prod.js",
     "analyze": "NODE_ENV=production webpack --config config/webpack.prod.js --profile --json > webpack.stats.json",
     "lint": "eslint  --ext .js --ext .jsx app/scripts config test",
     "lint:styles": "stylelint app/styles/**/*.scss",
-    "test": "jest --config config/.jestrc --coverage",
-    "test:watch": "jest --config config/.jestrc --watch",
+    "test": "jest --config config/jest.config.js --coverage",
+    "test:watch": "jest --config config/jest.config.js --watch",
     "test:ui": "nightwatch -c test/__setup__/nightwatch.conf.js",
     "test:ui:start": "npm start -- test:ui",
-    "deploy": "node config/tools deploy",
-    "publish": "node config/tools publish",
-    "docs": "node config/tools docs",
-    "precommit": "node config/tools commits && npm run lint && npm run lint:styles && npm test",
-    "postmerge": "node config/tools update && npm update"
+    "bundlesize": "bundlesize",
+    "deploy": "node tools deploy",
+    "publish": "node tools publish",
+    "docs": "node tools docs",
+    "precommit": "node tools commits && npm run lint && npm run lint:styles && npm test && npm run bundlesize",
+    "postmerge": "node tools update && npm update"
   },
-  "engines": {
-    "node": ">=6.0.0",
-    "npm": ">=3.0.0"
-  }
+  "browserslist": [
+    "ie >= 11",
+    "ie_mob >= 10",
+    "ff >= 30",
+    "chrome >= 34",
+    "safari >= 7",
+    "opera >= 23",
+    "ios >= 7",
+    "android >= 4.4",
+    "bb >= 10"
+  ],
+  "bundlesize": [
+    {
+      "path": "./dist/scripts/app.*.js",
+      "maxSize": "150 kB"
+    },
+    {
+      "path": "./dist/scripts/modernizr.*.js",
+      "maxSize": "30 kB"
+    }
+  ]
 }
