@@ -1,21 +1,65 @@
-import { persistStore } from 'redux-persist';
+import { applyMiddleware, createStore, compose, combineReducers } from 'redux';
+import thunk from 'redux-thunk';
+import createSagaMiddleware from 'redux-saga';
+import { routerReducer, routerMiddleware } from 'react-router-redux';
 
-let configStore;
+import { persistStore, persistReducer } from 'redux-persist';
+import storage from 'redux-persist/lib/storage';
 
-/* istanbul ignore else  */
-if (process.env.NODE_ENV === 'production' || process.env.NODE_ENV === 'test') {
-  configStore = require('./configureStore.prod').default;
-} else {
-  configStore = require('./configureStore.dev').default;
+import history from 'modules/history';
+import rootSagas from 'sagas';
+import rootReducer from 'reducers';
+
+const sagaMiddleware = createSagaMiddleware();
+
+const reducer = persistReducer(
+  {
+    key: 'septt', // key is required
+    storage, // storage is now required
+    whitelist: ['app', 'user'],
+  },
+  combineReducers({
+    ...rootReducer,
+    router: routerReducer,
+  })
+);
+
+
+const middleware = [
+  thunk,
+  sagaMiddleware,
+  routerMiddleware(history),
+];
+
+/* istanbul ignore next */
+if (process.env.NODE_ENV === 'development') {
+  const { createLogger } = require('redux-logger');
+
+  middleware.push(createLogger({ collapsed: true }));
 }
 
-const store = configStore();
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-/* istanbul ignore if  */
-if (process.env.NODE_ENV !== 'test') {
-  persistStore(store, {
-    whitelist: ['app', 'user']
-  });
-}
+/* istanbul ignore next */
+const configStore = (initialState = {}) => {
+  const createStoreWithMiddleware = composeEnhancers(applyMiddleware(...middleware))(createStore);
 
-export default store;
+  const store = createStoreWithMiddleware(reducer, initialState);
+
+  sagaMiddleware.run(rootSagas);
+
+  if (module.hot) {
+    module.hot.accept('reducers', () => {
+      store.replaceReducer(require('reducers').default);
+    });
+  }
+
+  return {
+    persistor: persistStore(store),
+    store,
+  };
+};
+
+const { store, persistor } = configStore();
+
+export { store, persistor };
