@@ -22,7 +22,7 @@ export function goTo(pathname, options = {}) {
   return push({
     pathname,
     search: options.search,
-    state: options.state
+    state: options.state,
   });
 }
 
@@ -40,7 +40,7 @@ export function showAlert(status, message, withTimeout = true) {
     type: ActionTypes.SHOW_ALERT,
     status,
     message,
-    withTimeout
+    withTimeout,
   };
 }
 
@@ -51,6 +51,6 @@ export function showAlert(status, message, withTimeout = true) {
  */
 export function hideAlert() {
   return {
-    type: ActionTypes.HIDE_ALERT
+    type: ActionTypes.HIDE_ALERT,
   };
 }
