@@ -1,4 +1,9 @@
+import Enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
 import 'vendor/polyfills';
+
+Enzyme.configure({ adapter: new Adapter() });
 
 Object.defineProperty(window.location, 'href', {
   writable: true,
@@ -15,18 +20,27 @@ Object.defineProperty(window.location, 'search', {
   value: '',
 });
 
+global.APP__TARGET = 'test';
+
 const react = document.createElement('div');
 react.id = 'react';
 react.style.height = '100vh';
 document.body.appendChild(react);
 
-window.__TARGET__ = 'test';
+const consoleError = console.error;
+console.error = jest.fn(message => {
+  const skipMessages = [
+    'redux-persist failed to create sync storage.',
+  ];
+  let shouldSkip = false;
 
-window.matchMedia = () =>
-  ({
-    matches: false,
-    addListener: () => {
-    },
-    removeListener: () => {
-    },
-  });
+  for (const s of skipMessages) {
+    if (message.includes(s)) {
+      shouldSkip = true;
+    }
+  }
+
+  if (!shouldSkip) {
+    consoleError(message);
+  }
+});
