@@ -11,7 +11,7 @@ describe('Footer', () => {
   const wrapper = setup();
 
   it('should be a StatelessComponent', () => {
-    expect(wrapper.instance().constructor.name).toBe('StatelessComponent');
+    expect(wrapper.instance()).toBeNull();
   });
 
   it('should render properly', () => {
