@@ -9,7 +9,7 @@ const mockDispatch = jest.fn();
 function setup(app = appState) {
   const props = {
     app,
-    dispatch: mockDispatch
+    dispatch: mockDispatch,
   };
 
   return mount(<SystemNotifications {...props} />);
@@ -19,7 +19,7 @@ describe('SystemNotifications', () => {
   const wrapper = setup();
 
   it('should be a StatelessComponent', () => {
-    expect(wrapper.instance().constructor.name).toBe('StatelessComponent');
+    expect(wrapper.instance()).toBeNull();
   });
 
   it('should render properly with the default state', () => {
@@ -33,9 +33,9 @@ describe('SystemNotifications', () => {
           status: 'success',
           message: 'Hello',
           visible: true,
-          withTimeout: true
-        }
-      }
+          withTimeout: true,
+        },
+      },
     });
 
     expect(wrapper.html()).toMatchSnapshot();
@@ -48,9 +48,9 @@ describe('SystemNotifications', () => {
           status: 'error',
           message: 'Fail',
           visible: true,
-          withTimeout: false
-        }
-      }
+          withTimeout: false,
+        },
+      },
     });
 
     expect(wrapper.html()).toMatchSnapshot();
@@ -68,9 +68,9 @@ describe('SystemNotifications', () => {
           status: 'error',
           message: 'Fail',
           visible: false,
-          withTimeout: false
-        }
-      }
+          withTimeout: false,
+        },
+      },
     });
     expect(wrapper.html()).toMatchSnapshot();
   });
