@@ -1,15 +1,21 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import App from 'containers/App';
+import { App } from 'containers/App';
 
-function setup() {
-  const props = {
-    children: <div key="1" className="child">Hello</div>,
-    location: {}
-  };
+const mockDispatch = jest.fn();
 
-  return shallow(<App {...props} />);
+const props = {
+  children: <div key="1" className="child">Hello</div>,
+  dispatch: mockDispatch,
+  location: {},
+  user: {
+    isAuthenticated: false,
+  },
+};
+
+function setup(ownProps = props) {
+  return shallow(<App {...ownProps} />);
 }
 
 describe('App', () => {
@@ -19,7 +25,22 @@ describe('App', () => {
     expect(wrapper.instance() instanceof React.Component).toBe(true);
   });
 
-  it('should render properly', () => {
-    expect(wrapper.html()).toMatchSnapshot();
+  it('should render properly for anonymous users', () => {
+    expect(wrapper.find('HelmetWrapper')).toBePresent();
+    expect(wrapper.find('ConnectedRouter')).toBePresent();
+    expect(wrapper.find('Switch')).toBePresent();
+    expect(wrapper.find('Footer')).toBePresent();
+    expect(wrapper.find('Connect(SystemNotifications)')).toBePresent();
+  });
+
+  it('should render properly for logged users', () => {
+    wrapper.setProps({
+      ...wrapper.props(),
+      user: {
+        isAuthenticated: true,
+      },
+    });
+
+    expect(wrapper.find('Header')).toBePresent();
   });
 });
